requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

fastapi==0.110.2
python-dotenv==1.0.1
uvicorn==0.29.0
asyncpg==0.29.0
sqlalchemy==2.0.29
pydantic==2.7.1
pyjwt==2.8.0